Split to multiple files for readability; standarize log messages; use structured logging
