+++ cmd/routesrv/polling.go
@@ -1,0 +1,120 @@
+package main
+
+import (
+	"context"
+	"fmt"
+	"sync"
+	"time"
+
+	ot "github.com/opentracing/opentracing-go"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+	log "github.com/sirupsen/logrus"
+	"github.com/zalando/skipper/routing"
+	"github.com/zalando/skipper/tracing"
+)
+
+const (
+	logPollingStarted       = "starting polling"
+	logPollingStopped       = "polling stopped"
+	logRoutesFetchingFailed = "failed to fetch routes"
+	logRoutesEmpty          = "received empty routes; ignoring"
+	logRoutesInitialized    = "routes initialized"
+	logRoutesUpdated        = "routes updated"
+)
+
+type pollerMetrics struct {
+	pollingStarted    prometheus.Gauge
+	routesInitialized prometheus.Gauge
+	routesUpdated     prometheus.Gauge
+}
+
+func newPollerMetrics() *pollerMetrics {
+	return &pollerMetrics{
+		pollingStarted: promauto.NewGauge(prometheus.GaugeOpts{
+			Namespace: "routesrv",
+			Name:      "polling_started_timestamp",
+			Help:      "UNIX time when the routes polling has started",
+		}),
+		routesInitialized: promauto.NewGauge(prometheus.GaugeOpts{
+			Namespace: "routesrv",
+			Name:      "routes_initialized_timestamp",
+			Help:      "UNIX time when the first routes were received and stored",
+		}),
+		routesUpdated: promauto.NewGauge(prometheus.GaugeOpts{
+			Namespace: "routesrv",
+			Name:      "routes_updated_timestamp",
+			Help:      "UNIX time of the last routes update (initial load counts as well)",
+		}),
+	}
+}
+
+type poller struct {
+	client  routing.DataClient
+	b       *eskipBytes
+	timeout time.Duration
+	quit    chan struct{}
+
+	tracer  ot.Tracer
+	metrics *pollerMetrics
+}
+
+func (p *poller) poll(wg *sync.WaitGroup) {
+	defer wg.Done()
+
+	var (
+		routesCount, routesBytes int
+		initialized              bool
+		msg                      string
+	)
+
+	log.WithFields(log.Fields{"timeout": p.timeout}).Info(logPollingStarted)
+	p.metrics.pollingStarted.SetToCurrentTime()
+	for {
+		span := tracing.CreateSpan("poll_routes", context.TODO(), p.tracer)
+
+		routes, err := p.client.LoadAll()
+		routesCount = len(routes)
+
+		switch {
+		case err != nil:
+			log.WithError(err).Error(logRoutesFetchingFailed)
+
+			span.SetTag("error", true)
+			span.LogKV(
+				"event", "error",
+				"message", fmt.Sprintf("%s: %s", logRoutesFetchingFailed, err),
+			)
+		case routesCount == 0:
+			log.Error(logRoutesEmpty)
+
+			span.SetTag("error", true)
+			span.LogKV(
+				"event", "error",
+				"message", msg,
+			)
+		case routesCount > 0:
+			routesBytes, initialized = p.b.formatAndSet(routes)
+			logger := log.WithFields(log.Fields{"count": routesCount, "bytes": routesBytes})
+			if initialized {
+				logger.Info(logRoutesInitialized)
+				span.SetTag("routes.initialized", true)
+				p.metrics.routesInitialized.SetToCurrentTime()
+			} else {
+				logger.Info(logRoutesUpdated)
+			}
+			p.metrics.routesUpdated.SetToCurrentTime()
+			span.SetTag("routes.count", routesCount)
+			span.SetTag("routes.bytes", routesBytes)
+		}
+
+		span.Finish()
+
+		select {
+		case <-p.quit:
+			log.Info(logPollingStopped)
+			return
+		case <-time.After(p.timeout):
+		}
+	}
+}
